PS-7044 feature: Provide Enterprise Encryption UDFs for OpenSSL (Digest operations)
https://jira.percona.com/browse/PS-7044

This commit should be squashed with commit 554745585e5
"PS-7044 feature: Provide Enterprise Encryption UDFs for OpenSSL (initial OpenSSL++)"

The following Digest operations functions put into a separate header
'opensslpp/digest_operations.hpp':
* calculate_digest()
