+++ extra/opensslpp/src/opensslpp/digest_operations.cpp
@@ -1,0 +1,29 @@
+/* Copyright (c) 2022 Percona LLC and/or its affiliates. All rights reserved.
+
+   This program is free software; you can redistribute it and/or
+   modify it under the terms of the GNU General Public License
+   as published by the Free Software Foundation; version 2 of
+   the License.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with this program; if not, write to the Free Software
+   Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA */
+
+#include "opensslpp/digest_operations.hpp"
+
+#include "opensslpp/digest_context.hpp"
+
+namespace opensslpp {
+
+std::string calculate_digest(const std::string &type, const std::string &data) {
+  digest_context ctx(type);
+  ctx.update(data);
+  return ctx.finalize();
+}
+
+}  // namespace opensslpp
