+++ plugin/encryption_udf/encryption_udf.cc
@@ -22,7 +22,7 @@
 #include <mysqlpp/udf_wrappers.hpp>
 
 #include <opensslpp/core_error.hpp>
-#include <opensslpp/digest_context.hpp>
+#include <opensslpp/digest_operations.hpp>
 #include <opensslpp/rsa_key.hpp>
 #include <opensslpp/rsa_operations.hpp>
 #include <opensslpp/rsa_padding.hpp>
